+++ src/ENDFtk/section/6/DiscreteTwoBodyScattering/TabulatedDistribution/test/TabulatedDistribution.test.cpp
@@ -223,6 +223,17 @@ void verifyChunkLANG12( const TabulatedDistribution& chunk ) {
   CHECK_THAT( 4., WithinRel( chunk.probabilities()[1] ) );
   CHECK_THAT( 6., WithinRel( chunk.probabilities()[2] ) );
 
+  CHECK( 1 == chunk.NR() );
+  CHECK( 1 == chunk.numberInterpolationRegions() );
+  CHECK( 1 == chunk.INT().size() );
+  CHECK( 1 == chunk.interpolants().size() );
+  CHECK( 1 == chunk.NBT().size() );
+  CHECK( 1 == chunk.boundaries().size() );
+  CHECK( 2 == chunk.INT()[0] );
+  CHECK( 2 == chunk.interpolants()[0] );
+  CHECK( 3 == chunk.NBT()[0] );
+  CHECK( 3 == chunk.boundaries()[0] );
+
   CHECK( 2 == chunk.NC() );
 }
 
@@ -257,6 +268,17 @@ void verifyChunkLANG14( const TabulatedDistribution& chunk ) {
   CHECK_THAT( 2., WithinRel( chunk.probabilities()[0] ) );
   CHECK_THAT( 4., WithinRel( chunk.probabilities()[1] ) );
   CHECK_THAT( 6., WithinRel( chunk.probabilities()[2] ) );
+
+  CHECK( 1 == chunk.NR() );
+  CHECK( 1 == chunk.numberInterpolationRegions() );
+  CHECK( 1 == chunk.INT().size() );
+  CHECK( 1 == chunk.interpolants().size() );
+  CHECK( 1 == chunk.NBT().size() );
+  CHECK( 1 == chunk.boundaries().size() );
+  CHECK( 4 == chunk.INT()[0] );
+  CHECK( 4 == chunk.interpolants()[0] );
+  CHECK( 3 == chunk.NBT()[0] );
+  CHECK( 3 == chunk.boundaries()[0] );
 
   CHECK( 2 == chunk.NC() );
 }